+++ src/BlogrCommentsPlugin.php
@@ -30,7 +30,7 @@ class BlogrCommentsPlugin implements BlogrExtension, FilamentPlugin
 
     public function getVersion(): string
     {
-        return '1.1.1';
+        return '1.1.3';
     }
 
     public function getAuthor(): string